+++ src/Service/JobRepository/BridgeChronos.php
@@ -12,6 +12,7 @@ namespace Chapi\Service\JobRepository;
 use Chapi\Component\Cache\CacheInterface;
 use Chapi\Component\Chronos\ApiClientInterface;
 use Chapi\Entity\Chronos\JobEntity;
+use Chapi\Service\JobValidator\JobValidatorServiceInterface;
 use Psr\Log\LoggerInterface;
 
 class BridgeChronos implements BridgeInterface
@@ -35,7 +36,7 @@ class BridgeChronos implements BridgeInterface
     private $oCache;
 
     /**
-     * @var JobEntityValidatorServiceInterface
+     * @var JobValidatorServiceInterface
      */
     private $oJobEntityValidatorService;
 
@@ -52,13 +53,13 @@ class BridgeChronos implements BridgeInterface
     /**
      * @param ApiClientInterface $oApiClient
      * @param CacheInterface $oCache
-     * @param JobEntityValidatorServiceInterface $oJobEntityValidatorService
+     * @param JobValidatorServiceInterface $oJobEntityValidatorService
      * @param LoggerInterface $oLogger
      */
     public function __construct(
         ApiClientInterface $oApiClient,
         CacheInterface $oCache,
-        JobEntityValidatorServiceInterface $oJobEntityValidatorService,
+        JobValidatorServiceInterface $oJobEntityValidatorService,
         LoggerInterface $oLogger
     )
     {