+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# dev-master (v0.8.x)
+# dev-master (v0.8.1)
+    2016-11-11 msiebeneicher <marc.siebeneicher@trivago.com>
+       * moved and refactored JobValidatorService
+       * improved `chapi validate` output
+       * added own command validation instead of simple `is not empty`
+
+# v0.8.0
     2016-11-04 msiebeneicher <marc.siebeneicher@trivago.com>
         * added container support
         * added constraints support